feat(desserbee-web): 재사용 위해 storeInfo 컴포넌트 분리
